SeriesOrder is a numeric(18,2) for sorting of 5a and 5b issues

+++ ComicsStore.Data/Model/BookSeries.cs
@@ -11,7 +11,7 @@ namespace ComicsStore.Data.Model
 
         public int BookId { get; set; }
         public string Issue { get; set; }
-        public int? SeriesOrder { get; set; }
+        public decimal? SeriesOrder { get; set; }
         public int SeriesId { get; set; }
 
         public Book Book { get; set; }
