+++ ComicsStore.MiddleWare/Models/Input/BookSeriesInputModel.cs
@@ -6,7 +6,7 @@ namespace ComicsStore.MiddleWare.Models.Input
     {
         public int SeriesId { get; set; }
         public string SeriesNr { get; set; }
-        public int? SeriesOrder { get; set; }
+        public decimal? SeriesOrder { get; set; }
 
         public SeriesInputModel Series { get; set; }
     }
