+++ ComicsStore.MiddleWare/Models/Output/SeriesBookOutputModel.cs
@@ -5,6 +5,6 @@ namespace ComicsStore.MiddleWare.Models.Output
     public class SeriesBookOutputModel : BasicBookOutputModel
     {
         public string SeriesNr { get; set; }
-        public int? SeriesOrder { get; set; }
+        public decimal? SeriesOrder { get; set; }
     }
 }
